Put Mac-specific code in its own file, raise error if platform unsupported
